fix: remove unreachable test
test_list_matched_files_missing_dirs is catching the error of the Mover instantiation rather than the ValueError thrown by list_matched_files. This error is unreachable because of the error handling of the Mover initialization.

A more comprehensive mock class would need to be set up in order to test the ValueError in list_matched_files.